Simplify data format
This commit simplifies the data format of the published data.
Originally this project was created with a somewhat overengineered and
difficult to handle format in mind, which made reasoning about the code
challenging und would not make much sense outside of its original
context. The new format is much easier, allowing also for the deletion
of two larger functions required for creating the old one.