Add timestamp and module index
This commits adds a timestamp to each dataset and includes an index
number to every module. Previously this information had been received
directly from the PLC, however this is highly redundant and not really
necessary. Instead of relying on the PLC providing those values, they
are now calculated directly within the program.